carbon-emissions-chatbot/requirements.txt
langchain 
openai 
psycopg2-binary 
langchain_experimental 
langchain-openai
gradio
python-dotenv